requirements.txt
alembic==1.4.3
arrow==0.17.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
parsedatetime==2.6
requests==2.18.4
SQLAlchemy==1.3.20
psycopg2==2.8.6
scancode-toolkit[full]
typecode-libmagic==5.39.210122
GitPython==3.1.13
shortuuid==1.0.1
APScheduler==3.6.3
python-whois==0.7.3
tldextract==3.1.0
python-nmap==0.6.1
python-owasp-zap-v2.4==0.0.18
lxml==4.6.2
beautifulsoup4==4.9.1
python-dateutil
